+++ lib/Views/splash_page.dart
